Revert "Send velocity as int and use pitch as uint8_t for transferring esp-now messages"
This reverts commit e19159f5ab6857bd1a027b8e4481c27ca598330f.

+++ guitar-midi-common/midi_common.h
@@ -23,7 +23,7 @@ enum Mode {
 struct ControllerMessage {
   DeviceType device;
   Mode mode;
-  uint8_t pitch;
+  int pitch;
   int velocity;
 };
